KDL type "branding" and include improvements
KDL files are now expected to contain, as their first non-blank
non-comment line, a node corresponding to their type. Currently this is
one of: app, board, or chip. In all three of these cases, that node
provides the name of the entity being described.

Inlined KDL is _not_ required to have this. Instead, the name is moved
to the node itself. So an inline board now looks like:

    board "board-name-here" {
        chip "foo"
    }

whereas an outlined board would read

    board "board-name-here"
    chip "foo"