Add Idyll enum bodies, use to fix report descriptors.
Technically, the host is permitted to request a prefix of the report
descriptor. In practice, every host I've tested always asks for 255
bytes (which in our case is enough), so I've been able to punt on this.
But, it was a good excuse to flesh out this Idyll feature.

The UsbEvent::ReportDescriptorNeeded variant now sports a 'length'
field, which carries the host's requested read length. The keyboard task
uses this to deliver a _prefix_ of the descriptor, if required.

This required adding enum fields to Idyll; they now appear to work.