tools: implement HUBRIS_IMAGE_ID properly
I hesitated to do this, because the image ID gets misused. It is _not_ a
unique build identifier, it is _not_ a git commit -- all it is is a
fingerprint with a fairly high chance of being different between two
firmwares. That's it.

But, Humility flash kind of relies on its presence, so, here we go. I've
attempted to incorporate many pieces of data that were omitted from the
original image ID, including the contents of all linker scripts.