idyll: restrict Message to Meta
Rather than passing the complete message descriptor from the kernel to
every idyll server method, we now only pass a restricted subset, the
Meta struct. This contains the sender and lease count for now.

This eliminates problems with aliasing the message buffer if we want to
provide a "tail" for variable length messages. (Though in that case we
may wind up wanting a reply size, too.)