Change path syntax to prepare for remote files.
The chip and board directives (in the board and app defs, respectively)
were referencing external definition files by "name." In practice, this
was being combined with a hardcoded path that made assumptions about
repository structure.

I don't really like forcing structure on other peoples' repositories
without good reason, and in this case there's a good reason _not_ to do
so: out-of-repository board and chipdefs are important, and previously
there was no obvious way to reference one.

I've made two changes here:

1. Boards and chips now need to be referenced by a path, including the
   file extension (.kdl typically).

2. The path needs to start with a prefix "proj:". This indicates that it
   should be handled relative to the project root, the place where
   hubake found hubris-env.toml.

The intent here is that other prefixes could be added, to support e.g.
pulling files from git or crates.io. But this is not implemented yet.

This may not be the final form of this feature, and feedback is of
course welcome.